move manager ringbuffer
