+++ manager.go
@@ -817,44 +817,6 @@ func (m *Manager) CloneMap(name string, newName string, options MapOptions) (*eb
 	return m.NewMap(spec, options)
 }
 
-// NewRingBuffer - Creates a new ring buffer and start listening for events.
-// Use a MapRoute to make this map available to the programs of the manager.
-func (m *Manager) NewRingBuffer(spec *ebpf.MapSpec, options MapOptions, ringBufferOptions RingBufferOptions) (*ebpf.Map, error) {
-	m.stateLock.Lock()
-	defer m.stateLock.Unlock()
-	if m.state < initialized {
-		return nil, ErrManagerNotInitialized
-	}
-
-	// check if the name of the new map is available
-	_, exists, _ := m.getMap(spec.Name)
-	if exists {
-		return nil, ErrMapNameInUse
-	}
-
-	// Create new map and ring buffer reader
-	ringBuffer, err := loadNewRingBuffer(spec, options, ringBufferOptions)
-	if err != nil {
-		return nil, err
-	}
-
-	// Setup ring buffer reader
-	if err := ringBuffer.init(m); err != nil {
-		return nil, err
-	}
-
-	// Start perf buffer reader
-	if err := ringBuffer.Start(); err != nil {
-		// clean up
-		_ = ringBuffer.Stop(CleanInternal)
-		return nil, err
-	}
-
-	// Add map to the list of perf ring managed by the manager
-	m.RingBuffers = append(m.RingBuffers, ringBuffer)
-	return ringBuffer.array, nil
-}
-
 // AddHook - Hook an existing program to a hook point. This is particularly useful when you need to trigger an
 // existing program on a hook point that is determined at runtime. For example, you might want to hook an existing
 // eBPF TC classifier to the newly created interface of a container. Make sure to specify a unique uid in the new probe,
